Moved ModelParameters to own module
